restructured and cleaned code
