+++ adrenaline.h
